move collaboration code into shared directory
