+++ package.json
@@ -27,10 +27,13 @@
   "dependencies": {
     "@types/dom-mediacapture-record": "^1.0.11",
     "@webaudiomodules/api": "webaudiomodules/api#commit=3ca912e8e08e66917eba22c330e85554932de494",
+    "lib0": "^0.2.47",
+    "monaco-editor": "^0.25.2",
     "preact": "^10.5.5"
   },
   "devDependencies": {
     "@webaudiomodules/sdk": "webaudiomodules/sdk#commit=810ba256f58385c0d50c16e36c82c7818bd01ead",
-    "@webaudiomodules/sdk-parammgr": "webaudiomodules/sdk-parammgr#commit=6eb3bc85fc30144502043f456f47268a62d3db5a"
+    "@webaudiomodules/sdk-parammgr": "webaudiomodules/sdk-parammgr#commit=6eb3bc85fc30144502043f456f47268a62d3db5a",
+    "wam-extensions": "boourns/wam-extensions#commit=7db0622f6b9e26389418ebadda61c2bba5938905"
   }
 }
