+++ src/plugins/functionseq/package.json
@@ -13,10 +13,7 @@
   "dependencies": {
     "@types/debug": "^4.1.6",
     "debug": "^4.3.1",
-    "lib0": "^0.2.47",
-    "monaco-editor": "^0.25.2",
-    "preact": "^10.5.5",
-    "wam-extensions": "boourns/wam-extensions#commit=7db0622f6b9e26389418ebadda61c2bba5938905"
+    "preact": "^10.5.5"
   },
   "devDependencies": {
     "copy-webpack-plugin": "^9.0.0",
